Refactor attestation imports 🛠️ Updated import paths for Attestation type to use new utility module.
